Add record to set CURLOPT_USERNAME (beware)
Using this along with to-be-added CURLOPT_PASSWORD would expose
credentials in the network.

I leave it in the driver as a means to completely express the API
functionalities, but there should be better ways of setting your
credentials if you mean to.

Probably will add a configuration file for that in the future.