Add curl file loading functionality
Add CurlLoadConfig record and loadConfigFile function.

loadConfigFile() calls writeInt32 and writeOctet for all parameters
it can recognize, except username and password. For those,
loadConfigFile() calls curl_easy_setopt without ever setting the
parameter value.

This means setpoint and RBV values for all parameters can be different,
so I add readback values for all records.

writeInt32 and writeOctet are called inside loadConfigFile(), which is
called inside writeInt32. In my understanding this cannot lead to any
infinite loop since CurlLoadConfig record has no "ASYN_" prepending it's
name, so it's impossible for it to call writeInt32 on itself.

Also, I think this is thread safe because the function is only supposed
to be called inside writeInt32, which is supposed to be already locked.